firmware/scripts/adsbee_util/pyproject.toml
[tool.poetry]
name = "adsbee-util"
version = "0.1.0"
description = ""
authors = ["John McNelly <jkailimcnelly@gmail.com>"]
readme = "README.md"
packages = [{include = "adsbee_util"}]

[tool.poetry.dependencies]
python = "^3.10"
pyqtgraph = "^0.13.4"
pyqt6 = "^6.6.1"
ipyleaflet = "^0.18.2"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
